Factor out lib/css.ts.
